demo-eng/adk_agents/ad_generation_agent/pyproject.toml
[project]
name = "ad_generation_agent"
version = "4.20260310.1"
description = "Product Catalog Ad Generation Agent"
authors = [{ name = "Gonzalo Varela", email = "gvarelal@google.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk==1.26.0",
    "google-cloud-aiplatform[adk,agent-engines]==1.140.0",
    "google-cloud-bigquery==3.40.1",
    "google-genai==1.66.0",
    "pydantic==2.12.5",
    "absl-py==2.4.0",
    "python-dotenv==1.2.2",
    "pandas==3.0.1",
    "toolbox-core==0.6.0",
    "moviepy==2.2.1",
    "google-cloud-texttospeech==2.34.0",
    "soundfile==0.13.1",
    "aiohttp==3.13.3",
    "adk_common",
    "requests==2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["ad_generation_agent"]
